Refatorando Mocks
