+++ tests/test_cnpj.py
@@ -25,33 +25,11 @@ class TestCNPJ(TestCase):
         )
         self.assertEqual(sieve("/...---.../"), "")
 
-    def test_remove_symbols(self):
-        with patch("brutils.cnpj.sieve") as mock_sieve:
-            # When call remove_symbols, it calls sieve
-            remove_symbols("12.345.678/0001-90")
-            mock_sieve.assert_called()
-
     def test_display(self):
         self.assertEqual(display("00000000000109"), "00.000.000/0001-09")
         self.assertIsNone(display("00000000000000"))
         self.assertIsNone(display("0000000000000"))
         self.assertIsNone(display("0000000000000a"))
-
-    def test_format_cnpj(self):
-        with patch("brutils.cnpj.is_valid", return_value=True) as mock_is_valid:
-            # When cnpj is_valid, returns formatted cnpj
-            self.assertEqual(
-                format_cnpj("01838723000127"), "01.838.723/0001-27"
-            )
-
-        # Checks if function is_valid_cnpj is called
-        mock_is_valid.assert_called_once_with("01838723000127")
-
-        with patch(
-            "brutils.cnpj.is_valid", return_value=False
-        ) as mock_is_valid:
-            # When cnpj isn't valid, returns None
-            self.assertIsNone(format_cnpj("01838723000127"))
 
     def test_validate(self):
         self.assertIs(validate("34665388000161"), True)
@@ -104,6 +82,34 @@ class TestCNPJ(TestCase):
         self.assertEqual(_checksum("00000000000000"), "00")
         self.assertEqual(_checksum("52513127000299"), "99")
 
+@patch("brutils.cnpj.sieve")
+class TestRemoveSymbols(TestCase):
+    def test_remove_symbols(self, mock_sieve):
+
+        # When call remove_symbols, it calls sieve
+        remove_symbols("12.345.678/0001-90")
+        mock_sieve.assert_called()
+
+@patch("brutils.cnpj.is_valid")
+class TestIsValidToFormat(TestCase):
+    def test_when_cnpj_is_valid_returns_true_to_format(self, mock_is_valid):
+        
+        mock_is_valid.return_value = True
+
+            # When cnpj is_valid, returns formatted cnpj
+        self.assertEqual(
+                format_cnpj("01838723000127"), "01.838.723/0001-27"
+            )
+
+        # Checks if function is_valid_cnpj is called
+        mock_is_valid.assert_called_once_with("01838723000127")
+
+    def test_when_cnpj_is_not_valid_returns_none(self, mock_is_valid):
+        
+        mock_is_valid.return_value = False
+        
+        # When cnpj isn't valid, returns None
+        self.assertIsNone(format_cnpj("01838723000127"))
 
 if __name__ == "__main__":
     main()
